refactor(hud): strip local-player bars, rename HUD.ts to EnemyHealthBars.ts
